# Improve performance of WSGI server under high traffic