--- a/environments/python/Dockerfile
+++ b/environments/python/Dockerfile
@@ -1,7 +1,7 @@
 FROM alpine:3.5
 
 RUN apk update
-RUN apk add --no-cache python3 python3-dev build-base libev-dev
+RUN apk add --no-cache python3 python3-dev build-base
 RUN pip3 install --upgrade pip
 RUN rm -r /root/.cache
 
@@ -11,4 +11,3 @@ RUN pip3 install -r requirements.txt
 
 ENTRYPOINT ["python3"]
 CMD ["server.py"]
-