--- a/environments/python/Dockerfile-2.7
+++ b/environments/python/Dockerfile-2.7
@@ -1,7 +1,7 @@
 FROM alpine:3.5
 
 RUN apk update
-RUN apk add --no-cache python python-dev build-base py-pip libev-dev
+RUN apk add --no-cache python python-dev build-base py-pip
 RUN pip install --upgrade pip
 RUN rm -r /root/.cache
 
@@ -11,4 +11,3 @@ RUN pip install -r requirements.txt
 
 ENTRYPOINT ["python"]
 CMD ["server.py"]
-