--- a/environments/python/requirements.txt
+++ b/environments/python/requirements.txt
@@ -1,5 +1,5 @@
 Flask===0.11.1
-bjoern
+gevent
 httplib2
 python-dateutil
 requests==2.7.0