@soamvasani 2018-05-21
Thanks, @xiekeyang !  I can definitely see the big improvement in the 99%le time.  But it does seem to come at the expense of median (47ms vs 189ms) time, as well as at the expense of throughput (2041 vs 1282 req/s).  Even the 98%le time increases.