@xiekeyang 2018-05-23
@soamvasani , Thanks, I noticed this. IMO, it should be most considered that the response times of `gevnet` are much more well-distributed than `bjoern`. 

```
  50%    189
  66%    197
  75%    215
  80%    223
  90%    244
  95%    265
  98%    275
  99%    282
 100%   1225 (longest request)
```

**vs**

```
  50%     47
  66%     86
  75%     97
  80%    100
  90%    108
  95%    124
  98%    157
  99%    3815
  100%   3827 (longest request)
```

I think this is because that `bjoern` is a little weak on multi-thread, and it is strong and compact design on WSGI, so it handles first requests so quickly. Meantime `gevent` set up multi-thread cache, and assign requests to different threads.

After long time, `bjoern` reach to its capacity limit and requests have to wait. Meantime, `gevent` has free space to handle them. So it shows `well distributed`.

What's more, my paste above is simple script printing "hello". I tested some higher load script:

```py
import numpy as np
import pandas as pd

def func1(weight, feature, intercept):
    linear_result = sum(weight * feature) + intercept
    prob = 1 / (1 + np.exp(-1 * linear_result))
    b = - (100 / np.log(2))
    a = round(100 + b * np.log(0.1), 0)
    return a - b * prob

def func2(args):
    trans_df = pd.DataFrame({"code": data_code, "standard": data_standard, "woe": data_woe})

    feature_input = []
    none_num = 0
    for feature in features:
        feature_value = args[feature]
        if feature_value is None:
            none_num += 1
        feature_input.append(trans_to_score(trans_df, feature, args[feature]))
    if none_num == len(features):
        return None
    return feature_input
```

The report likes:

### bjoern

The http server is closed because of high load.

```
$ ab -c 256 -n 8192 http://127.0.0.1:8888/
This is ApacheBench, Version 2.3 <$Revision: 1706008 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient)
Completed 819 requests
Completed 1638 requests
Completed 2457 requests
Completed 3276 requests
Completed 4095 requests
Completed 4914 requests
apr_socket_recv: Connection reset by peer (104)
Total of 5637 requests completed
```

### gevent

It can finish without `Failed` request and keep well distribution.

```
$ ab -c 256 -n 8192 http://127.0.0.1:8888/
This is ApacheBench, Version 2.3 <$Revision: 1706008 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient)
Completed 819 requests
Completed 1638 requests
Completed 2457 requests
Completed 3276 requests
Completed 4095 requests
Completed 4914 requests
Completed 5733 requests
Completed 6552 requests
Completed 7371 requests
Completed 8190 requests
Finished 8192 requests


Server Software:        
Server Hostname:        127.0.0.1
Server Port:            8888

Document Path:          /
Document Length:        17 bytes

Concurrency Level:      256
Time taken for tests:   180.330 seconds
Complete requests:      8192
Failed requests:        0
Total transferred:      1245184 bytes
HTML transferred:       139264 bytes
Requests per second:    45.43 [#/sec] (mean)
Time per request:       5635.324 [ms] (mean)
Time per request:       22.013 [ms] (mean, across all concurrent requests)
Transfer rate:          6.74 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   1.4      0      14
Processing:    37 5558 1369.8   5342    8531
Waiting:       37 5557 1369.7   5342    8531
Total:         46 5558 1369.2   5342    8531

Percentage of the requests served within a certain time (ms)
  50%   5342
  66%   6346
  75%   6750
  80%   6899
  90%   7308
  95%   7744
  98%   7966
  99%   8143
 100%   8531 (longest request)
```

After I deploy it on K8S, I observe `gevent` pods and find memory and CPU figure reach high level quickly under high traffic. It is actually better on resource utilization.