@xiekeyang 2018-05-24
According to well-distributed time, served within shorter time, and be able to support more concurrent requests, I prefer to `gevent`. Please reconsider it. Thanks a lot!