@life1347 2018-05-24
Before the result is out, here some thoughts:

1. If the PR tries to improve RPS, then we may need to focus on `Requests per second` since it's not fair to compare two frameworks' 99%le response time that under different RPS.

2. Gevent use multi-thread to serve traffic and increases resourcec utilization, it's ideal if we launch one gevent pod. But for a "function" it's really not a good idea to serve such high load traffic with only one function pod.  In fission's case, newdeploy as function exectortype brings ability of auto-scaling for function. With newdeploy and bjoern, we can have short response time, high rps and also high resource utilization.