@soamvasani 2018-06-08
I don't think we have to choose between the two.  I can certainly imagine that there is a use case for a gevent-based environment.  I think the default python environment should remain bjoern, for the median/95%le performance, and for the reasons @life1347 describes above. 

We can have a separate python-gevent directory in the environments, and add it to the build.