@xiekeyang 2018-06-08
> We can have a separate python-gevent directory in the environments, and add it to the build.

@soamvasani I think it is a good idea! I would set up another PR to add `gevent` env.